CartSummary moved to cart.py
